Bugfix while checking

+++ Outcomes/outcomes.py
@@ -29,12 +29,12 @@ def run_delayframe_outcomes(config, setup_name, outdir, scenario_seir, scenario_
             parameters[new_comp]['source'] = config_outcomes[new_comp]['source'].as_str()
             parameters[new_comp]['probability'] = np.mean(
                 config_outcomes[new_comp]['probability']['value'].as_random_distribution()(size = 10000))
-            parameters[new_comp]['delay'] = np.round(np.mean(
-                config_outcomes[new_comp]['delay']['value'].as_random_distribution()(size = 10000)))
+            parameters[new_comp]['delay'] = int(np.round(np.mean(
+                config_outcomes[new_comp]['delay']['value'].as_random_distribution()(size = 10000))))
             
             if config_outcomes[new_comp]['duration'].exists():
-                parameters[new_comp]['duration'] = np.round(np.mean(
-                    config_outcomes[new_comp]['duration']['value'].as_random_distribution()(size = 10000)))
+                parameters[new_comp]['duration'] = int(np.round(np.mean(
+                    config_outcomes[new_comp]['duration']['value'].as_random_distribution()(size = 10000))))
             
             if (config["outcomes"]["param_from_file"].get()):
                 # stuff to put it in parameter table
